% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard.R
\name{ggtree_jaccard}
\alias{ggtree_jaccard}
\title{Generate dendrogram with Jaccard distance and clusters}
\usage{
ggtree_jaccard(data, nclust_jaccard = 2, offset_label = 0.2)
}
\arguments{
\item{data}{Data frame containing the SSR marker data}

\item{`nclust_jaccard`}{Number of clusters to divide the dendrogram into}

\item{`offset_label`}{Offset factor for label positioning in ggtree plot}
}
\value{
A list with the following elements:
\itemize{
\item Jaccard distance matrix
\item Jaccard distance matrix as gt table
\item Cluster assignments from cutree
\item Cluster assignments as gt table
\item ggtree plot colored by cluster
}
}
\description{
This function calculates the Jaccard distance matrix, clusters the data
using hierarchical clustering, cuts the dendrogram into a specified
number of clusters, and generates a ggtree fan dendrogram colored by
cluster.
}
\examples{

library(pbgeno)
  
data(Jaccard_example)  
  
ggtree_out <- ggtree_jaccard(Jaccard_example, nclust_jaccard = 2)
 
}
