% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jaccard_example.R
\docType{data}
\name{Jaccard_example}
\alias{Jaccard_example}
\title{Example Jaccard distance matrix dataset}
\format{
An object of class \code{data.frame} with 58 rows and 42 columns.
}
\usage{
Jaccard_example
}
\description{
Simulated genotype calls for 41 genotypes across 21 markers,
to demonstrate Jaccard distance matrix calculation
}
\keyword{datasets}
