% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Example dataset of marker metadata}
\format{
An object of class \code{data.frame} with 174 rows and 11 columns.
}
\usage{
metadata
}
\description{
Contains marker ID (rs), chromosome and position information for
for markers that are in the Structure_example data
}
\keyword{datasets}
