% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic_calc_fun.R
\name{pic_calc}
\alias{pic_calc}
\title{Calculate PIC values for molecular marker data}
\usage{
pic_calc(data)
}
\arguments{
\item{data}{Data frame with marker genotype data. First column should
be the marker names. Subsequent columns are the genotypes for each
sample.}
}
\value{
A list with the following components:
\itemize{
\item pic: Data frame with marker name, heterozygosity, and PIC value
\item data: Original input data frame
}
}
\description{
This function calculates polymorphism information content (PIC)
values for each marker in the given data set. It handles both
co-dominant and dominant markers appropriately.
}
\examples{
data(PIC_example)
results <- pic_calc(PIC_example)

}
