% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIC_example.R
\docType{data}
\name{PIC_example}
\alias{PIC_example}
\title{Example dataset for calculating PIC values}
\format{
An object of class \code{data.frame} with 6 rows and 9 columns.
}
\usage{
PIC_example
}
\description{
Genotype data for 2 SSR markers scored across 7 wheat lines.
First column is marker name and additional columns are genotypes.
}
\keyword{datasets}
