% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Str2tassel.R
\name{str2tassel}
\alias{str2tassel}
\title{Convert STRUCTURE format data to TASSEL format}
\usage{
str2tassel(data, met, na.strings = "-9")
}
\arguments{
\item{data}{Data frame with genotype calls and marker names}

\item{met}{The meta data for the tassel data format}

\item{na.strings}{Character strings specifying missing genotype calls}
}
\value{
Data frame in TASSEL format with metadata and genotype calls
}
\description{
This function converts STRUCTURE format genotype data format to the
format required by TASSEL software. It handles recoding alleles,
imputing missing data, pairing alleles, and attaching marker metadata. The
}
\examples{
 
data(Structure_example)
tassel_data <- str2tassel(data=Structure_example,met=metadata,na.strings="-9")
write.table(tassel_data,"Tassel_hapmap.txt", row.names = F,quote = F, sep = "\t")
write.csv(res_f,"Tassel_hapmap.csv", row.names = F,quote = F)

  
}
