% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_example.R
\docType{data}
\name{Structure_example}
\alias{Structure_example}
\title{Example dataset for Structure format}
\format{
An object of class \code{data.frame} with 240 rows and 175 columns.
}
\usage{
Structure_example
}
\description{
This dataset contains genotype calls for 240 lines across
174 SSR markers, formatted for input into Structure software
}
\keyword{datasets}
